requirements.txt
yfinance>=0.2.40
pandas>=2.0
pandas-ta>=0.3.14b
python-telegram-bot>=20.0
